+++ src/txmempool.h
@@ -46,12 +46,12 @@ private:
     int64_t nTime;             //!< Local time when entering the mempool
     double dPriority;          //!< Priority when entering the mempool
     unsigned int nHeight;      //!< Chain height when entering the mempool
-    unsigned int sigOpCost;    //!< Total sigop cost
+    int64_t sigOpCost;    //!< Total sigop cost
 
 public:
     CTxMemPoolEntry(const CTransaction& _tx, const CAmount& _nFee,
                     int64_t _nTime, double _dPriority, unsigned int _nHeight,
-                    unsigned int nSigOpsCost);
+                    int64_t nSigOpsCost);
     CTxMemPoolEntry(const CTransaction& _tx, const CAmount& _nFee, int64_t _nTime, double _dPriority, unsigned int _nHeight);
     CTxMemPoolEntry(const CTxMemPoolEntry& other);
     CTxMemPoolEntry();
@@ -62,7 +62,7 @@ public:
     size_t GetTxSize() const { return nTxSize; }
     int64_t GetTime() const { return nTime; }
     unsigned int GetHeight() const { return nHeight; }
-    unsigned int GetSigOpCost() const { return sigOpCost; }
+    int64_t GetSigOpCost() const { return sigOpCost; }
 };
 
 class CMinerPolicyEstimator;
